@@ -64,7 +64,7 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <release>15</release>
+                    <release>16</release>
                 </configuration>
             </plugin>
             <plugin>
