@@ -27,7 +27,7 @@ public class onPlayerItemConsume extends SimpleListener implements Listener, Eve
         	//e.getPlayer().getInventory().getItemInMainHand().setAmount(e.getPlayer().getInventory().getItemInMainHand().getAmount() - 1);
         	Bukkit.getServer().getScheduler().scheduleSyncDelayedTask(getPlugin(), () -> {e.getPlayer().getInventory().setItemInMainHand(null);});
         	if (((PotionMeta)e.getItem().getItemMeta()).hasCustomEffect(PotionEffectType.INVISIBILITY)) {
-        		getPlugin().getPlayers().get(e.getPlayer().getUniqueId()).hide();
+        		getPlugin().getPlayers().get(e.getPlayer().getUniqueId()).hideArmor();
         	}
         	
             if(((PotionMeta) e.getItem().getItemMeta()).getBasePotionData().getType().equals(PotionType.INVISIBILITY)) {
