@@ -8,6 +8,7 @@ import main.Config;
 import main.Plugin;
 
 import java.util.Collection;
+import java.util.UUID;
 
 public class PartyManager{
 
@@ -15,7 +16,7 @@ public class PartyManager{
         for(Team team : plugin.getTeams().values()){
             if(team.getTeammatesAmount() == Config.getPlayersPerTeam()) continue;
             if(checkPlayerInTeam(party, team.getTeammates().values())){
-                TeamManager.addPlayerToTeam(plugin, team, plugin.getPlayers().get(player.uuid));
+                plugin.getPlayers().get(UUID.fromString(player.uuid)).setTeam(team);
                 return;
             }
         }
@@ -25,18 +26,18 @@ public class PartyManager{
     private static void addToEmptyTeam(Plugin plugin, Party party, Player player){
         for(Team team : plugin.getTeams().values()){
             if(team.getTeammatesAmount() != 0) continue;
-            TeamManager.addPlayerToTeam(plugin, team, plugin.getPlayers().get(player.uuid));
+            plugin.getPlayers().get(UUID.fromString(player.uuid)).setTeam(team);
             return;
         }
-        addToRandomTeam(plugin, party, player);
+        addToRandomTeam(plugin, player);
     }
 
-    private static void addToRandomTeam(Plugin plugin, Party party, Player player){
+    private static void addToRandomTeam(Plugin plugin, Player player){
         String teamColor = Config.getTeamsNames().get(0);
         for(Team team : plugin.getTeams().values())
             if(plugin.getTeams().get(teamColor).getTeammatesAmount() > team.getTeammatesAmount())
                 teamColor = team.getColor();
-        TeamManager.addPlayerToTeam(plugin, plugin.getTeams().get(teamColor), plugin.getPlayers().get(player.uuid));
+        plugin.getPlayers().get(UUID.fromString(player.uuid)).setTeam(plugin.getTeams().get(teamColor));
     }
 
     private static boolean checkPlayerInTeam(Party party, Collection<Participant> participants){
